requirements.txt
django==3.2.13
django-guardian==2.4.0
djangorestframework==3.13.1
